moving few method from cronService->pointService
